backend/requirements.txt
flask 
google-generativeai 
python-dotenv 
requests 
duckduckgo-search
flask_cors
ddgs